Point SSE wind at 158 degrees, not 150
WIND_DIRECTIONS maps IMS's wind direction ids to degrees for the wind
direction sensor and the weather entity's wind bearing. Id 8 (SSE) was 150,
off the 22.5-degree compass rose every other entry follows; IMS's own table
says 158. Inherited from upstream, which has the same value.

The test checks every id against the captured wind_directions response, so
a future drift in either shows up. Verified to fail against the previous
const.py.

+++ custom_components/ims_motinium/const.py
@@ -307,7 +307,7 @@ WIND_DIRECTIONS = {
     5: float(90),
     6: float(113),
     7: float(135),
-    8: float(150),
+    8: float(158),
     9: float(180),
     10: float(203),
     11: float(225),
